4.20-4.21 done

+++ controllers/blogs.js
@@ -3,14 +3,6 @@ const Blog = require("../models/blog")
 const User = require("../models/user")
 const jwt= require("jsonwebtoken")
 
-const getTokenFrom = request => {
-    const authorization = request.get('authorization')
-    if (authorization && authorization.startsWith('Bearer ')) {
-        return authorization.replace('Bearer ', '')
-    }
-    return null
-    }
-
 blogsRouter.get('/', async (request, response) => {
     const blogs = await Blog.find({}).populate("user", {username: 1, name: 1})
     response.json(blogs)
@@ -27,8 +19,7 @@ blogsRouter.get("/:id", async (request, response) => {
 
 blogsRouter.post('/', async (request, response) => {
     const body = request.body
-
-    const decodedToken = jwt.verify(getTokenFrom(request), process.env.SECRET)
+    const decodedToken = jwt.verify(request.token, process.env.SECRET)
     if (!decodedToken.id) {
         return response.status(401).json({ error: 'token invalid' })
     }
@@ -61,9 +52,18 @@ blogsRouter.post('/', async (request, response) => {
 })
 
 blogsRouter.delete("/:id", async (request, response) => {
-    await Blog.findByIdAndDelete(request.params.id)
-    //const user = await User.findById(body.userId)
-    response.status(204).end()
+    const decodedToken = jwt.verify(request.token, process.env.SECRET)
+    if (!decodedToken.id) {
+        return response.status(401).json({ error: 'token invalid' })
+    }
+    const user = await User.findById(decodedToken.id)
+    const blog = await Blog.findById(request.params.id)
+    if (blog.user.toString() == user._id.toString()) {
+        await Blog.findByIdAndDelete(request.params.id)
+        response.status(204).end()
+    } else {
+        response.status(400).json({error : "You can only delete your own blogs"})
+    }
 })
 
 blogsRouter.put("/:id", async (request, response) => {
